Add AIO ALB
The AIO control option requires some means of communicating with the
cluster during setup.  This is best handled by an ALB, since there is
an ASG that needs to be referenced.  While it is expected that the ALB
will generally only be accessed during cluster bootstrapping, it is
handy to have a direct line that does not depend on any Nomad
supervised proxies.